Updated vehicle controller
- Removed seek destination method
- Removed seek method
- Removed stop method
