+++ Assets/New Folder/Scripts/VehicleFleet.cs
@@ -16,26 +16,6 @@ public class VehicleFleet : MonoBehaviour
     public ClockHour Hr;
     public ClockMinute Min;
 
-    private void Awake()
-    {
-        clock.MinutePassed += Chekk;
-    }
-
-    private void Chekk()
-    {
-        if (clock.IsSameTimeOfDay(Hr, Min))
-        {
-            StartCoroutine(Rout());
-        }
-    }
-
-    private IEnumerator Rout()
-    {
-        vehicles[0].SeekDestination(Seek);
-        yield return new WaitForSeconds(3);
-        vehicles[0].SeekDestination(Seek2);
-    }
-
     public int GetVehicleCount()
     {
         return vehicles.Count;
